requirements.txt
Flask>=2.0
requests>=2.28
beautifulsoup4>=4.12
python-dateutil>=2.8
scapy>=2.5.0   # optional; used for advanced fingerprinting (run as root)
dnspython>=2.2.0
